requirements.txt
absl-py==0.12.0
appdirs==1.4.4
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
Click==7.0
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
distlib==0.3.1
distro==1.4.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.3.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
matplotlib==3.3.4
mccabe==0.6.1
more-itertools==4.2.0
mysql==0.0.3
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.3
netifaces==0.10.4
numpy==1.20.1
oauthlib==3.1.0
pbr==5.5.1
pexpect==4.6.0
Pillow==8.1.0
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
SoundFile==0.10.3.post1
ssh-import-id==5.10
stevedore==3.3.0
testresources==2.0.1
Twisted==18.9.0
urllib3==1.25.8
virtualenv==20.4.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
Werkzeug==1.0.1
zipp==1.0.0
zope.interface==4.7.1
